% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gampls.R
\name{gampls}
\alias{gampls}
\title{GAM-PLS model using MGCV}
\usage{
gampls(fit, large = FALSE, ...)
}
\arguments{
\item{fit}{.}

\item{large}{Boolean, default to FALSE. If `large` is TRUE then `mgcv::bam` rather
than `mgcv::gam` is used.}

\item{...}{.}
}
\value{
.
}
\description{
fit_gampls
}
\author{
Pierre Roudier
}
