% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret.R
\docType{data}
\name{gamplsInfo}
\alias{gamplsInfo}
\title{a GAM-PLS model info to use in caret::train}
\format{
An object of class \code{list} of length 9.
}
\usage{
gamplsInfo
}
\description{
gamplsInfo
}
\examples{
\dontrun{

library(spectacles)
library(caret)
oz <- load_oz()

fit <- train(
  x = spectra(oz),
  y = oz$carbon,
  method = gamplsInfo,
  tuneGrid = expand.grid(
    ncomp = 3:10,
    select = FALSE,
    method = "GCV.Cp"
  ),
  trControl = trainControl(
    method = "repeatedcv",
    number = 10,
    repeats = 5,
    # Here we can specify the summary function used during parametrisation
    summaryFunction = spectroSummary
 ),
  # Here we can specify which metric to use to optimise the model parameters
 metric = "RPIQ"
)

}

}
\author{
Pierre Roudier
}
\keyword{datasets}
