% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gampls}
\alias{predict.gampls}
\title{Prediction from a fitted GAM-PLS model}
\usage{
\method{predict}{gampls}(object, newdata, type = "response", ...)
}
\arguments{
\item{object}{a fitted GAM-PLS model, typically an output of \code{gampls}}

\item{newdata}{A \code{data.frame} or list containing the values of the model covariates at which predictions are required.}

\item{type}{See \code{mgcv::predict.gam}, defaults to \code{"response"}}

\item{...}{Further arguments passed to \code{mgcv::predict.gam}}
}
\value{
See \code{mgcv::predict.gam}
}
\description{
predict
}
\details{
This is a simple wrapper around \code{mgcv::predict.gam}.
}
\author{
Pierre Roudier
}
