% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{summarise_preds}
\alias{summarise_preds}
\title{Summarise predictions from a GAM-PLS}
\usage{
summarise_preds(sim, probs = c(0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975), ...)
}
\arguments{
\item{sim}{A \code{data.frame} of the posterior samples, generated using \code{simulate}}

\item{probs}{A numeric vector of probabilities with values in \code{[0, 1]}, passed to \code{stats::quantile}}

\item{...}{Further arguments passed to \code{inlabru::bru_summarise}}
}
\value{
A \code{data.frame} with summary statistics, \code{mean}, \code{sd}, the quantiles as per the \code{probs} argument, and the MC statistics \code{mean.mc_std_err} and \code{sd.mc_std_err}
}
\description{
summarise_preds
}
\details{
The function is a simple wrapper around \code{inlabru::bru_summarise}
}
\author{
Pierre Roudier
}
