% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{simulate}
\alias{simulate}
\title{Simulate from the posterior distribution using {gratia}}
\usage{
simulate(fit, n, data, ...)
}
\arguments{
\item{fit}{A GAM fitted on the latent variables of a PLS}

\item{n}{numeric; the number of posterior simulations to return}

\item{data}{data.frame; observations for which the posterior draws from the models should be evaluated.}

\item{...}{Fyurther arguments passed to \code{gratia::simulate.gam}}
}
\value{
A matrix with \code{nsim} columns.
}
\description{
Simulations from the posterior distribution of the GAM fitted from the latent variables of the original PLS regression model. This uses \code{gratia::simulate.gam}.
}
\details{
This is a simple wrapper around \code{gratia::simulate.gam}, see the dedicated help page.
}
\author{
Pierre Roudier
}
